move files out of commands folder
