+++ tools/code-analyzer/src/print.ts
