+++ tools/code-analyzer/src/utils.ts
